i32

+++ utils.h
@@ -60,7 +60,7 @@ void utils_error_mode(i32 code);
 
 void utils_internal_led(bool on);
 
-u16 utils_proportional_reduce(u16 number, u16 step, u16 total_steps);
+i32 utils_proportional_reduce(i32 number, i32 step, i32 total_steps);
 
 i32 utils_scaled_pwm_percentage(i32 val, i32 deadzone, i32 max_val);
